stockanalyser_agent/pyproject.toml
[project]
name = "stock-analyser-agent"
version = "0.1.0"
description = "Stock analyser agent for the A2A project."
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk>=0.2.5",
    "google-adk>=1.2.1",
    "google-genai",
    "python-dotenv",
    "uvicorn",
    "requests",
] 
